# swagger.json
{
  "swagger": "2.0",
  "info": {
    "description": "@openapi:info",
    "version": "0.0.1"
  },
  "paths": {
...
    "/api/chemistryprobes": {
      "get": {
        "description": "List all chemistry probes in reef-pi.",
        "tags": [
          "ChemistryProbes"
        ],
        "summary": "List all chemistry probes.",
        "operationId": "chemistryProbeList",
        "responses": {
          "200": {
            "description": "phProbe",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/phProbe"
              }
            }
          }
        }
      },
      "put": {
        "description": "Create a new chemistry probe.",
        "tags": [
          "ChemistryProbes"
        ],
        "summary": "Create a chemistry probe.",
        "operationId": "chemistryProbeCreate",
        "parameters": [
          {
            "description": "The chemistry probe to create",
            "name": "phProbe",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/phProbe"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/chemistryprobes/{id}": {
      "get": {
        "description": "Get an existing chemistry probe by id.",
        "tags": [
          "ChemistryProbes"
        ],
        "summary": "Get a chemistry probe by id.",
        "operationId": "chemistryProbeGet",
        "parameters": [
          {
            "description": "The Id of the chemistry probe",
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/phProbe"
            }
          },
          "404": {
            "description": "Not Found"
          }
        }
      },
      "post": {
        "description": "Update an existing chemistry probe.",
        "tags": [
          "ChemistryProbes"
        ],
        "summary": "Update a chemistry probe.",
        "operationId": "chemistryProbeUpdate",
        "parameters": [
          {
            "description": "The Id of the chemistry probe to update",
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            }
          },
          {
            "description": "The chemistry probe to update",
            "name": "phProbe",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/phProbe"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "404": {
            "description": "Not Found"
          }
        }
      },
      "delete": {
        "description": "Delete an existing chemistry probe.",
        "tags": [
          "ChemistryProbes"
        ],
        "summary": "Delete a chemistry probe.",
        "operationId": "chemistryProbeDelete",
        "parameters": [
          {
            "description": "The Id of the chemistry probe to delete",
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/chemistryprobes/{id}/calibrate": {
      "post": {
        "description": "Set calibration points for one or two point calibration",
        "tags": [
          "ChemistryProbes"
        ],
        "summary": "Calibrate a chemistry probe.",
        "operationId": "chemistryProbeCalibrate",
        "parameters": [
          {
            "description": "The Id of the chemistry probe to calibrate",
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            }
          },
          {
            "description": "The calibration measurements",
            "name": "measurements",
            "in": "body",
            "required": true,
            "schema": {
              "type": "array",
              "items": {
                "type": "object",
                "properties": {
                  "expected": {
                    "description": "The expected value",
                    "type": "float64",
                    "format": "float"
                  },
                  "observed": {
                    "description": "The actual value observed",
                    "type": "float64",
                    "format": "float"
                  }
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/chemistryprobes/{id}/calibratepoint": {
      "post": {
        "description": "Set a calibration point for one or two point calibration",
        "tags": [
          "ChemistryProbes"
        ],
        "summary": "Calibrate a chemistry probe.",
        "operationId": "chemistryProbeCalibrateSingle",
        "parameters": [
          {
            "description": "The Id of the chemistry probe to calibrate",
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            }
          },
          {
            "description": "The calibration measurement",
            "name": "measurements",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/calibrationPoint"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/chemistryprobes/{id}/read": {
      "get": {
        "description": "Get chemistry probe reading.",
        "tags": [
          "ChemistryProbes"
        ],
        "summary": "Get chemistry probe reading.",
        "operationId": "chemistryProbeRead",
        "parameters": [
          {
            "description": "The Id of the chemistry probe to read",
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/chemistryprobes/{id}/readings": {
      "get": {
        "description": "Get chemistry probe readings.",
        "tags": [
          "ChemistryProbes"
        ],
        "summary": "Get chemistry probe readings.",
        "operationId": "chemistryProbeReadings",
        "parameters": [
          {
            "description": "The Id of the chemistry probe to delete",
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/statsResponse"
            }
          }
        }
      }
    },
...
    "/api/phprobes": {
      "get": {
        "description": "List all ph probes in reef-pi.",
        "tags": [
          "PhProbes"
        ],
        "summary": "List all ph probes.",
        "operationId": "phProbeList",
        "responses": {
          "200": {
            "description": "phProbe",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/phProbe"
              }
            }
          }
        }
      },
      "put": {
        "description": "Create a new ph probe.",
        "tags": [
          "PhProbes"
        ],
        "summary": "Create a ph probe.",
        "operationId": "phProbeCreate",
        "parameters": [
          {
            "description": "The ph probe to create",
            "name": "phProbe",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/phProbe"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/phprobes/{id}": {
      "get": {
        "description": "Get an existing ph probe by id.",
        "tags": [
          "PhProbes"
        ],
        "summary": "Get an ph probe by id.",
        "operationId": "phProbeGet",
        "parameters": [
          {
            "description": "The Id of the ph probe",
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/phProbe"
            }
          },
          "404": {
            "description": "Not Found"
          }
        }
      },
      "post": {
        "description": "Update an existing ph probe.",
        "tags": [
          "PhProbes"
        ],
        "summary": "Update a ph probe.",
        "operationId": "phProbeUpdate",
        "parameters": [
          {
            "description": "The Id of the ph probe to update",
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            }
          },
          {
            "description": "The ph probe to update",
            "name": "phProbe",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/phProbe"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "404": {
            "description": "Not Found"
          }
        }
      },
      "delete": {
        "description": "Delete an existing ph probe.",
        "tags": [
          "PhProbes"
        ],
        "summary": "Delete a ph probe.",
        "operationId": "phProbeDelete",
        "parameters": [
          {
            "description": "The Id of the ph probe to delete",
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/phprobes/{id}/calibrate": {
      "post": {
        "description": "Set calibration points for one or two point calibration",
        "tags": [
          "PhProbes"
        ],
        "summary": "Calibrate a ph probe.",
        "operationId": "phProbeCalibrate",
        "parameters": [
          {
            "description": "The Id of the ph probe to calibrate",
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            }
          },
          {
            "description": "The calibration measurements",
            "name": "measurements",
            "in": "body",
            "required": true,
            "schema": {
              "type": "array",
              "items": {
                "type": "object",
                "properties": {
                  "expected": {
                    "description": "The expected value",
                    "type": "float64",
                    "format": "float"
                  },
                  "observed": {
                    "description": "The actual value observed",
                    "type": "float64",
                    "format": "float"
                  }
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/phprobes/{id}/calibratepoint": {
      "post": {
        "description": "Set a calibration point for one or two point calibration",
        "tags": [
          "PhProbes"
        ],
        "summary": "Calibrate a ph probe.",
        "operationId": "phProbeCalibrateSingle",
        "parameters": [
          {
            "description": "The Id of the ph probe to calibrate",
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            }
          },
          {
            "description": "The calibration measurement",
            "name": "measurements",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/calibrationPoint"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/phprobes/{id}/read": {
      "get": {
        "description": "Get ph probe reading.",
        "tags": [
          "PhProbes"
        ],
        "summary": "Get ph probe reading.",
        "operationId": "phProbeRead",
        "parameters": [
          {
            "description": "The Id of the ph probe to read",
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/phprobes/{id}/readings": {
      "get": {
        "description": "Get ph probe readings.",
        "tags": [
          "PhProbes"
        ],
        "summary": "Get ph probe readings.",
        "operationId": "phProbeReadings",
        "parameters": [
          {
            "description": "The Id of the ph probe to delete",
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/statsResponse"
            }
          }
        }
      }
    },
...
  },
  "definitions": {
...
    "ChartConfig": {
      "type": "object",
      "properties": {
        "color": {
          "type": "string",
          "x-go-name": "Color"
        },
        "ymax": {
          "type": "number",
          "format": "double",
          "x-go-name": "YMax"
        },
        "ymin": {
          "type": "number",
          "format": "double",
          "x-go-name": "YMin"
        }
      },
      "x-go-package": "github.com/reef-pi/reef-pi/controller/modules/temperature"
    },
    "Duration": {
      "description": "A Duration represents the elapsed time between two instants\nas an int64 nanosecond count. The representation limits the\nlargest representable duration to approximately 290 years.",
      "type": "integer",
      "format": "int64",
      "x-go-package": "time"
    },
...
    "Metric": {
      "type": "object",
      "x-go-package": "github.com/reef-pi/reef-pi/controller/telemetry"
    },
...
    "Notify": {
      "type": "object",
      "properties": {
        "enable": {
          "type": "boolean",
          "x-go-name": "Enable"
        },
        "max": {
          "type": "number",
          "format": "double",
          "x-go-name": "Max"
        },
        "min": {
          "type": "number",
          "format": "double",
          "x-go-name": "Min"
        }
      },
      "x-go-package": "github.com/reef-pi/reef-pi/controller/modules/temperature"
    },
...
    "calibrationPoint": {
      "type": "object",
      "properties": {
        "expected": {
          "type": "number",
          "format": "double",
          "x-go-name": "Expected"
        },
        "observed": {
          "type": "number",
          "format": "double",
          "x-go-name": "Observed"
        },
        "type": {
          "type": "string",
          "x-go-name": "Type"
        }
      },
      "x-go-name": "CalibrationPoint",
      "x-go-package": "github.com/reef-pi/reef-pi/controller/modules/chemistry"
    },
...
    "phProbe": {
      "type": "object",
      "properties": {
        "analog_input": {
          "type": "string",
          "x-go-name": "AnalogInput"
        },
        "chart": {
          "$ref": "#/definitions/ChartConfig"
        },
        "control": {
          "type": "boolean",
          "x-go-name": "Control"
        },
        "downer_eq": {
          "type": "string",
          "x-go-name": "DownerEq"
        },
        "enable": {
          "type": "boolean",
          "x-go-name": "Enable"
        },
        "hysteresis": {
          "type": "number",
          "format": "double",
          "x-go-name": "Hysteresis"
        },
        "id": {
          "type": "string",
          "x-go-name": "ID"
        },
        "is_macro": {
          "type": "boolean",
          "x-go-name": "IsMacro"
        },
        "max": {
          "type": "number",
          "format": "double",
          "x-go-name": "Max"
        },
        "min": {
          "type": "number",
          "format": "double",
          "x-go-name": "Min"
        },
        "name": {
          "type": "string",
          "x-go-name": "Name"
        },
        "notify": {
          "$ref": "#/definitions/Notify"
        },
        "one_shot": {
          "type": "boolean",
          "x-go-name": "OneShot"
        },
        "period": {
          "$ref": "#/definitions/Duration"
        },
        "temp_sensor_id": {
          "type": "integer",
          "format": "int64",
          "x-go-name": "TempSensorID"
        },
        "transformer": {
          "type": "string",
          "x-go-name": "Transformer"
        },
        "upper_eq": {
          "type": "string",
          "x-go-name": "UpperEq"
        }
      },
      "x-go-name": "Probe",
      "x-go-package": "github.com/reef-pi/reef-pi/controller/modules/chemistry"
    },
...
    "statsResponse": {
      "type": "object",
      "properties": {
        "current": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/Metric"
          },
          "x-go-name": "Current"
        },
        "historical": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/Metric"
          },
          "x-go-name": "Historical"
        }
      },
      "x-go-name": "StatsResponse",
      "x-go-package": "github.com/reef-pi/reef-pi/controller/telemetry"
    },
...
  }
}